Create feb1trnsn
All submissions for this problem are available.Akash has an array of size N. Akash gets really triggered whenever he sees arrays with too many transitions, so he wants to rearrange the array so that the number of transitions becomes minimum possible.

The number of transitions of an array is the number of indices such that when you move to the adjacent index (towards the right), the value at the new index is different from the value at the previous index. For example the number of transitions for [1,1,2,4] is 2 as there are only 2 transition positions (indices 2 and 3).