Create xorit
You are given an array of n integers. Then you are given q queries.Each query contains two integers L and R , and you have to find the XOR of all the elements in between L and R inclusive.
Print q lines , each line containing an integer which is the XOR of all elements from index L to R inclusive.